requirements.txt
numpy
opencv-python
tqdm
Flask
python-dotenv
gunicorn